// src/selectors/kpi/models/IKpiSettingTableRow.ts
// export default interface IKpiSettingTableRow {
//     readonly kpiID: string;
//     readonly keyActivity: string;
//     readonly divisionNameList: string;
//     readonly departmentNameList: string;
//     readonly kpiDireksiName: string;
//     readonly measurementName: string;
//     readonly weight: number;
//     readonly point: number;
//     readonly statusName: string;
// };

export default interface IKpiSettingTableRow {
    readonly kpiID: string;
    readonly keyActivity: string;
    readonly divisionNameList: string;
    readonly departmentNameList: string;
    readonly functionNameList: string;
    readonly kpiDireksiName: string;
    readonly measurementName: string;
    readonly weight: number;
    readonly point: number;
    readonly statusName: string;
};